// src/app/models/invitation/candidateInvitationViewModel.ts
export class CandidateInvitationViewModel {
  invitationId:number;
  candidateId: string;
  candidateCode: number;
  firstName: string;
  lastName: string;
  fatherName: string;
  provinceTitle: string;
  electionZoneTitle: string;
  invitationDate: string;
  invitationStatus: string;
  invitationDescription: string;
}